List entries in a directory of a virtual ADF device

Get an overview of all entries (files and directories) in a specific directory.

Usage

adf_dir_list(dev, virtual_path, recursive = FALSE, ...)

# S3 method for class 'adf_device'
adf_dir_list(dev, virtual_path, recursive = FALSE, ...)

# S3 method for class 'character'
adf_dir_list.adf_device(dev, virtual_path, recursive = FALSE, ...)

# S3 method for class 'virtual_path'
adf_dir_list.adf_device(dev, virtual_path, recursive = FALSE, ...)

Arguments

dev

The virtual adf device for which information needs to be obtained. It should be of class adf_device which can be created with create_adf_device() or connect_adf().

virtual_path

The virtual path for which you wish to obtain a list of entries (see also vignette("virtual_paths")). When missing, entries for the current directory (adf_directory()) are returned.

recursive

A logical value. When set to TRUE, the function is called recursively for all subdirectories in virtual_path.

...

Ignored

Value

A vector of virtual_path class objects

Author

Pepijn de Vries

Examples

## First setup a connection to a virtual device
adz_file <- system.file("example.adz", package = "adfExplorer2")
my_device <- connect_adf(adz_file)

## List all entries in the disk's root:
adf_dir_list(my_device)
#> DIR  ...-------- Devs
#> DIR  ...   -------- S
#> DIR  ...-------- this
#> DIR  ...-------- mods
## List all entries on the disk:
adf_dir_list(my_device, recursive = TRUE)
#> DIR  ...-------- Devs
#> FILE ...configuration
#> DIR  ...   -------- S
#> FILE ...rtup-Sequence
#> DIR  ...-------- this
#> DIR  ...  -------- is
#> DIR  ...   -------- a
#> DIR  ...-------- deep
#> DIR  ...-------- path
#> DIR  ...-------- mods
#> FILE ...--- mod.intro

close(my_device)